Улучшения по стилю кодирования
- В стиле кодирования от Microsoft для С++ классы имеют префикс C, а их файлы - нет
- Применил constructor initializer lists в конструкторе CBankClient
- Убрал циклический #include между Bank.h и BankClient.h, который не будет работать в рамках препроцессора и раздельной компиляции C/C++

+++ PP2/Bank.cpp
@@ -1,4 +1,4 @@
-#include "CBank.h"
+#include "Bank.h"
 
 CBank::CBank()
 {
@@ -9,7 +9,7 @@ CBank::CBank()
 
 CBankClient* CBank::CreateClient()
 {
-	unsigned int clientId = m_clients.size();
+	unsigned clientId = unsigned(m_clients.size());
 	CBankClient* client = new CBankClient(this, clientId);
 	m_clients.push_back(*client);
 	return client;
